# server/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorizationParams": {
        "additionalProperties": false,
        "properties": {
          "client_id": {
            "type": "string"
          },
          "code_challenge": {
            "type": "string"
          },
          "code_challenge_method": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "response_type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "redirect_uri",
          "response_type",
          "code_challenge",
          "code_challenge_method"
        ],
        "type": "object"
      },
...
      "ConnectedApp": {
        "additionalProperties": false,
        "properties": {
          "clientId": {
            "type": "string"
          },
          "created": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastUsed": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientId",
          "name",
          "scope",
          "created",
          "lastUsed"
        ],
        "type": "object"
      },
      "ConsentPrompt": {
        "additionalProperties": false,
        "properties": {
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string"
          },
          "loopback": {
            "description": "The answer goes to a program on this computer, e.g. Claude Code.",
            "type": "boolean"
          },
          "redirectHost": {
            "description": "Where the answer goes. Shown so a person can spot an impostor.",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "clientName",
          "clientId",
          "redirectHost",
          "loopback",
          "scopes"
        ],
        "type": "object"
      },
      "Decide-consentResponse": {
        "additionalProperties": false,
        "properties": {
          "redirectTo": {
            "type": "string"
          }
        },
        "required": [
          "redirectTo"
        ],
        "type": "object"
      },
      "Decision": {
        "additionalProperties": false,
        "properties": {
          "approve": {
            "type": "boolean"
          },
          "params": {
            "$ref": "#/components/schemas/AuthorizationParams"
          }
        },
        "required": [
          "params",
          "approve"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "examples": [
              "not_found"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemDetail"
            },
            "type": "array"
          },
          "status": {
            "examples": [
              404
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "examples": [
              "Not Found"
            ],
            "type": "string"
          },
          "type": {
            "examples": [
              "about:blank"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "An OAuth access token, as issued to MCP clients.",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "description": "The web session cookie set by signing in with a passkey.",
        "in": "cookie",
        "name": "__Host-plonkout_sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Workout logging and training plans.",
    "title": "Plonkout API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/account/connected-apps": {
      "get": {
        "operationId": "list-connected-apps",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectedApp"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Apps the user has let into their account, e.g. Claude",
        "tags": [
          "oauth"
        ]
      }
    },
    "/api/account/connected-apps/{id}": {
      "delete": {
        "operationId": "disconnect-app",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an app's access",
        "tags": [
          "oauth"
        ]
      }
    },
...
    "/api/oauth/consent": {
      "get": {
        "operationId": "get-consent-prompt",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "client_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "redirect_uri",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "response_type",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "code_challenge",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "code_challenge_method",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "scope",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "resource",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentPrompt"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Describe an authorization request for the consent page",
        "tags": [
          "oauth"
        ]
      },
      "post": {
        "description": "Answers with where to send the browser: back to the client with a code, or with access_denied.",
        "operationId": "decide-consent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decide-consentResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Approve or deny an authorization request",
        "tags": [
          "oauth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}